AMD as factory function
Looking at https://requirejs.org/docs/whyamd.html#definition a factory
function should be registered instead of immediate execution.

This solves an issue I'm having when importing qz-tray in ember. I do the
following:

    app.import('node_modules/qz-tray/qz-tray.js', {
      using: [
        { transformation: 'amd', as: 'qz-tray' }
      ]
    });

See https://guides.emberjs.com/release/addons-and-dependencies/managing-dependencies/#toc_anonymous-amd-javascript-modules

Without the change in this PR I get an error when the ember app is
loading, it expects a function that is will execute and not an object.

BUT https://github.com/amdjs/amdjs-api/blob/master/AMD.md#factory-
specifies that it can be a function OR object.